+++ mrmustard/math/lattice/strategies/julia/compactFock/helperFunctions.jl
