+++ mrmustard/math/lattice/strategies/julia/compactFock/singleLeftoverMode_amps.jl
