+++ mrmustard/math/lattice/strategies/julia/compactFock/singleLeftoverMode_grad.jl
